+++ src/Sentry.Unity/SentryUnitySDK.cs
@@ -99,4 +99,18 @@ internal class SentryUnitySdk
                 "Exception while releasing the lockfile on the config directory.", ex);
         }
     }
+
+    public SentryUnity.CrashedLastRun CrashedLastRun()
+    {
+        if (_options.CrashedLastRun is null)
+        {
+            _options.DiagnosticLogger?.LogDebug("The SDK does not have a 'CrashedLastRun' set. " +
+                                                "This might be due to a missing or disabled native integration.");
+            return SentryUnity.CrashedLastRun.Unknown;
+        }
+
+        return _options.CrashedLastRun.Invoke()
+            ? SentryUnity.CrashedLastRun.Crashed
+            : SentryUnity.CrashedLastRun.DidNotCrash;
+    }
 }